Move Reminder.java and ReminderTest.java back to reminder directory Fix clock before calling Clock.now() in ReminderTest
