+++ src/main/java/seedu/momentum/logic/parser/ParserUtil.java
@@ -16,7 +16,7 @@ import seedu.momentum.logic.parser.exceptions.ParseException;
 import seedu.momentum.model.project.Deadline;
 import seedu.momentum.model.project.Description;
 import seedu.momentum.model.project.Name;
-import seedu.momentum.model.project.Reminder;
+import seedu.momentum.model.reminder.Reminder;
 import seedu.momentum.model.tag.Tag;
 
 /**
