+++ src/main/java/seedu/momentum/model/project/Project.java
@@ -7,6 +7,7 @@ import java.util.Set;
 
 import javafx.collections.ObservableList;
 import seedu.momentum.commons.core.DateWrapper;
+import seedu.momentum.model.reminder.Reminder;
 import seedu.momentum.model.reminder.ReminderManager;
 import seedu.momentum.model.tag.Tag;
 import seedu.momentum.model.timer.TimerWrapper;
