+++ src/test/java/seedu/momentum/logic/parser/ParserUtilTest.java
@@ -19,7 +19,7 @@ import seedu.momentum.logic.parser.exceptions.ParseException;
 import seedu.momentum.model.project.Deadline;
 import seedu.momentum.model.project.Description;
 import seedu.momentum.model.project.Name;
-import seedu.momentum.model.project.Reminder;
+import seedu.momentum.model.reminder.Reminder;
 import seedu.momentum.model.tag.Tag;
 
 public class ParserUtilTest {
